// frontend/src/app/model/Categorie.ts
export class Categorie {
    //Categ_id: String;
    categ_numero: number;
    categ_numerobranche: number;
    categ_libellelong: String;
    categ_libellecourt: String;
    categ_classificationanalytique: String;
    categ_codetaxe: number;
    categ_codecommission: number;
    categ_datemodification: Date;
    categ_datecomptabilisation: Date;
}