// frontend/src/app/model/Commission.ts
export class Commission {

    comm_id: number ;
    comm_code: number ;
	comm_codeapporteur: number ;
    comm_codebranche: number ;
    comm_codecategorie: number ;
    comm_codeproduit: number ;
    comm_codegarantie: number ;
    comm_typecalcul: string ;
    comm_interv1: number ;
    comm_interv2: number ;
    comm_interv3: number ;
    comm_interv4: number ;
    comm_interv5: number ;
    comm_interv6: number ;
    comm_interv7: number ;
    comm_interv8: number ;
    comm_interv9: number ;
    comm_interv10: number ;
    comm_tauxcommission12: number ;
    comm_montantforfait12: number ;
    comm_tauxcommission34: number ;
    comm_montantforfait34: number ;
    comm_tauxcommission56: number ;
    comm_montantforfait56: number ;
    comm_tauxcommission78: number ;
    comm_montantforfait78: number ;
    comm_tauxcommission910: number ;
    comm_montantforfait910: number ;
    comm_datepriseffet: Date ;
    comm_datefineffet: Date ;
}